aaaaahhhh all broken

+++ bitset.c
@@ -3,6 +3,6 @@
 extern inline void bitset_free(bitset_index_t *name);
 extern inline unsigned long bitset_size(bitset_index_t *name);
 extern inline void bitset_fill(bitset_index_t *name, bool set);
-extern inline void bitset_setbit(bitset_index_t *name, unsigned index, bool set);
-extern inline unsigned long bitset_getbit(bitset_index_t *name, unsigned index);
+extern inline void bitset_setbit(bitset_index_t *name, unsigned long index, bool set);
+extern inline unsigned long bitset_getbit(bitset_index_t *name, unsigned long index);
 #endif
